// src/types/editor.ts
export type ItemType = "text" | "image" | "video" | "audio" | "helper" | "caption" | "template" | "composition" | "illustration" | "shape" | "rect" | "progressBar" | "progressFrame" | "progressSquare" | "radialAudioBars" | "linealAudioBars" | "waveAudioBars" | "hillAudioBars";
// ...
export interface ISize {
    width: number;
    height: number;
    type?: string;
    name?: string;
}
export interface IDisplay {
    from: number;
    to: number;
}
export interface ITrim {
    from: number;
    to: number;
}
export interface IBoxShadow {
    color: string;
    x: number;
    y: number;
    blur: number;
}
export interface IMetadata {
    resourceId: string;
    order: number;
}
export interface IBasicAnimation {
    name: string;
    composition: ICompositionAnimation[];
}
export interface ICompositionAnimation {
    property: string;
    from: number;
    to: number;
    durationInFrames: number;
    ease?: (t: number) => number;
    easing: string;
    delay: number;
    details?: {
        fonts: {
            fontFamily: string;
            url: string;
        }[];
    };
    text?: string;
    persist?: boolean;
}
export interface ITransition {
    id: string;
    trackId: string;
    fromId: string;
    toId: string;
    type: string;
    name?: string;
    duration: number;
    preview?: string;
    direction?: any;
    kind: string;
}
export interface ITrack {
    id: string;
    type: ITrackType;
    items: string[];
    muted?: boolean;
    metadata?: Partial<IMetadata>;
    accepts?: string[];
    index?: number;
    magnetic?: boolean;
    static?: boolean;
}
export interface ITimelineScaleState {
    unit: number;
    zoom: number;
    segments: number;
    index: number;
}
export interface ITimelineScrollState {
    /**
     * Timeline scroll state by X-axis.
     */
    left: number;
    /**
     * Timeline scroll state by Y-axis.
     */
    top: number;
}
export interface CanvasSpacing {
    left: number;
    right: number;
}
export interface ItemStructure {
    id: string;
    items: string[];
    transitions: string[];
    tracks: ITrack[];
}
interface ICommonDetails {
    width?: number;
    height?: number;
    transform?: string;
    opacity?: number;
    border?: string;
    borderRadius?: number;
    boxShadow?: IBoxShadow;
    top?: number | string;
    left?: number | string;
    borderColor?: string;
    borderWidth?: number;
}
export interface ITrackItemBase {
    id: string;
    name: string;
    type: ItemType;
    preview?: string;
    display: IDisplay;
    duration?: number;
    trim?: ITrim;
    isMain?: boolean;
    animations?: {
        in: IBasicAnimation;
        out: IBasicAnimation;
        loop: IBasicAnimation;
        timed: IBasicAnimation;
    };
    playbackRate?: number;
    modifier?: IDisplay;
    details?: any;
    activeEdit?: boolean;
    metadata: Record<string, any>;
    transitionInfo?: {
        isFrom: boolean;
        isTo: boolean;
        transition: ITransition;
    };
}
export type ITrackItem = IAudio | IImage | IText | IVideo | ICaption | ITemplate | IComposition | IIllustration | IShape | IRect | IProgressBar | IProgressSquare | IProgressFrame | IRadialAudioBars | ILinealAudioBars | IWaveAudioBars | IHillAudioBars;
export interface ICaptionWord {
    end: number;
    start: number;
    word: string;
}
export interface ICaptionDetails extends ICommonDetails {
    skewX: number;
    skewY: number;
    text: string;
    fontSize: number;
    fontFamily: string;
    fontUrl: string;
    color: string;
    lineHeight: number | string;
    letterSpacing: number | string;
    fontWeight: number;
    fontStyle: string;
    textDecoration: string;
    textAlign: "center" | "left" | "right";
    wordSpacing: number | string;
    textShadow: string;
    backgroundColor: string;
    opacity: number;
    width: number;
    height: number;
    top: number | string;
    left: number | string;
    border: string;
    wordWrap: "normal" | "break-word";
    wordBreak: "normal" | "break-word" | "break-all";
    WebkitTextStrokeColor: string;
    WebkitTextStrokeWidth: string;
    borderWidth: number;
    borderColor: string;
    boxShadow: {
        color: string;
        x: number;
        y: number;
        blur: number;
    };
    textTransform: "capitalize" | "uppercase" | "lowercase";
    words: ICaptionWord[];
    appearedColor?: string;
    activeColor?: string;
    activeFillColor?: string;
    animation?: string;
    isKeywordColor?: string;
    preservedColorKeyWord?: boolean;
    linesPerCaption?: number;
    wordsPerLine?: string;
    showObject?: string;
}
export interface ITextDetails extends ICommonDetails {
    skewX: number;
    skewY: number;
    text: string;
    fontSize: number;
    fontFamily: string;
    fontUrl: string;
    color: string;
    lineHeight: number | string;
    letterSpacing: number | string;
    fontWeight: number;
    fontStyle: string;
    textDecoration: string;
    textAlign: "center" | "left" | "right";
    wordSpacing: number | string;
    textShadow: string;
    backgroundColor: string;
    opacity: number;
    width: number;
    textTransform: "capitalize" | "uppercase" | "lowercase";
    height: number;
    top: number | string;
    left: number | string;
    border: string;
    wordWrap: "normal" | "break-word";
    wordBreak: "normal" | "break-word" | "break-all";
    WebkitTextStrokeColor: string;
    WebkitTextStrokeWidth: string;
    borderWidth: number;
    borderColor: string;
    boxShadow: {
        color: string;
        x: number;
        y: number;
        blur: number;
    };
}
export interface IImageDetails extends ICommonDetails {
    src: string;
    background: string;
    width: number;
    height: number;
    opacity: number;
    transform: string;
    border: string;
    borderRadius: number;
    boxShadow: IBoxShadow;
    top: string;
    left: string;
    transformOrigin: string;
    crop: {
        x: number;
        y: number;
        width: number;
        height: number;
    };
    blur: number;
    brightness: number;
    flipX: boolean;
    flipY: boolean;
    rotate: string;
    visibility: "visible" | "hidden";
}
export interface IVideoDetails extends ICommonDetails {
    src: string;
    frames?: number;
    background?: string;
    stream?: ReadableStream<Uint8Array>;
    blob?: Blob;
    width: number;
    height: number;
    volume?: number;
    boxShadow?: IBoxShadow;
    transformOrigin?: string;
    crop?: {
        x: number;
        y: number;
        width: number;
        height: number;
    };
    blur: number;
    brightness: number;
    flipX: boolean;
    flipY: boolean;
    rotate: string;
    visibility: "visible" | "hidden";
}
export interface IAudioDetails extends ICommonDetails {
    src: string;
    volume?: number;
}
export interface IIllustrationDetails extends ICommonDetails {
    path: string;
    src: string;
    width: number;
    height: number;
    opacity: number;
    transform: string;
    border: string;
    top: string;
    left: string;
    flipX: boolean;
    flipY: boolean;
    rotate: string;
    svgString: string;
    initialSvgString: string;
    visibility: "visible" | "hidden";
    colorMap: Record<string, string>;
}
export interface IShapeDetails extends ICommonDetails {
    path: string;
    src: string;
    width: number;
    height: number;
    opacity: number;
    transform: string;
    border: string;
    top: string;
    left: string;
    flipX: boolean;
    flipY: boolean;
    rotate: string;
    visibility: "visible" | "hidden";
    backgroundColor: string;
}
export interface IRectDetails extends ICommonDetails {
    width: number;
    height: number;
    opacity: number;
    transform: string;
    border: string;
    top: string;
    left: string;
    flipX: boolean;
    flipY: boolean;
    rotate: string;
    visibility: "visible" | "hidden";
    backgroundColor: string;
    boxShadow: IBoxShadow;
    blur: number;
    brightness: number;
}
export interface IProgressBarDetails extends ICommonDetails {
    width: number;
    height: number;
    top: string;
    left: string;
    backgroundColors: string[];
    border: string;
    borderRadius: number;
    borderWidth: number;
    borderColor: string;
    opacity: number;
    flipX: boolean;
    flipY: boolean;
    inverted: boolean;
}
export interface IProgressFrameDetails extends ICommonDetails {
    width: number;
    height: number;
    top: string;
    left: string;
    backgroundColors: string[];
    border: string;
    borderRadius: number;
    borderWidth: number;
    borderColor: string;
    opacity: number;
    flipX: boolean;
    flipY: boolean;
    inverted: boolean;
    barThickness: number;
}
export interface IProgressSquareDetails extends ICommonDetails {
    width: number;
    height: number;
    top: string;
    left: string;
    strokeColors: string[];
    strokeWidth: number;
    inverted: boolean;
    strokeBackground: string;
    flipX: boolean;
    flipY: boolean;
    rotate: string;
    transform: string;
    opacity: number;
}
export interface IRadialAudioBarsDetails extends ICommonDetails {
    width: number;
    height: number;
    top: string;
    left: string;
    border: string;
    borderRadius: number;
    borderWidth: number;
    borderColor: string;
    opacity: number;
    flipX: boolean;
    flipY: boolean;
    srcs: string[];
    reproduceAudio: boolean;
    radialBarColor: string;
    audioDatas: any[];
}
export interface ILinealAudioBarsDetails extends ICommonDetails {
    width: number;
    height: number;
    top: string;
    left: string;
    border: string;
    borderRadius: number;
    borderWidth: number;
    borderColor: string;
    opacity: number;
    flipX: boolean;
    flipY: boolean;
    srcs: string[];
    reproduceAudio: boolean;
    audioDatas: any[];
    linealBarColor: string;
    backgroundColor: string;
    lineThickness: number;
    gapSize: number;
    roundness: number;
    placement: string | null;
    inverted: boolean;
    strokeColor: string;
    fillColor: string | null;
    strokeWidth: number | null;
    copies: number | null;
    offsetPixelSpeed: number;
    lineColor: string | string[];
    lineGap: number;
    topRoundness: number;
    bottomRoundness: number;
    lines: number;
    sections: number;
}
// ...
export interface IText extends ITrackItemBase {
    type: "text";
    details: ITextDetails;
}
export interface ICaption extends ITrackItemBase {
    type: "caption";
    details: ICaptionDetails;
}
export interface IImage extends ITrackItemBase {
    type: "image";
    details: IImageDetails;
}
export interface IVideo extends ITrackItemBase {
    type: "video";
    details: IVideoDetails;
}
export interface IAudio extends ITrackItemBase {
    type: "audio";
    details: IAudioDetails;
}
export interface IIllustration extends ITrackItemBase {
    type: "illustration";
    details: IIllustrationDetails;
}
export interface IShape extends ITrackItemBase {
    type: "shape";
    details: IShapeDetails;
}
export interface IRect extends ITrackItemBase {
    type: "rect";
    details: IRectDetails;
}
export interface IProgressBar extends ITrackItemBase {
    type: "progressBar";
    details: IProgressBarDetails;
}
export interface IProgressSquare extends ITrackItemBase {
    type: "progressSquare";
    details: IProgressSquareDetails;
}
export interface IProgressFrame extends ITrackItemBase {
    type: "progressFrame";
    details: IProgressFrameDetails;
}
export interface IRadialAudioBars extends ITrackItemBase {
    type: "radialAudioBars";
    details: IRadialAudioBarsDetails;
}
export interface ILinealAudioBars extends ITrackItemBase {
    type: "linealAudioBars";
    details: ILinealAudioBarsDetails;
}
export interface IWaveAudioBars extends ITrackItemBase {
    type: "waveAudioBars";
    details: IWaveAudioBarsDetails;
}
export interface IHillAudioBars extends ITrackItemBase {
    type: "hillAudioBars";
    details: IHillAudioBarsDetails;
}
export interface IComposition extends ITrackItemBase {
    type: "composition";
    trackItemIds: string[];
    trackItemsMap: Record<string, ITrackItem>;
    tracks: ITrack[];
    size: ISize;
}
export interface ITemplate extends ITrackItemBase {
    type: "template";
    trackItemIds: string[];
    trackItemsMap: Record<string, ITrackItem>;
    transitionsMap: Record<string, ITransition>;
    transitionIds: string[];
    size: ISize;
    tracks: ITrack[];
    structure: ItemStructure[];
}
// ...
export type ITrackItemAndDetails = ITrackItem;